Deletes unused setup.rs
